Update title and favicon in various views
This commit updates the title and favicon in various views of the application. The favicon has been changed from the default Vite icon to the CLIMBS logo. The title has been updated to reflect the name of the application or the specific page being displayed. The updated title and favicon provide a more consistent and branded experience for the user.